Rename dir to remove tsx extension
